# Light optimization on last activities loading