--- a/decidim-core/app/cells/decidim/activities_cell.rb
+++ b/decidim-core/app/cells/decidim/activities_cell.rb
@@ -27,13 +27,13 @@ module Decidim
     end
 
     def activities
-      @activities ||= last_activities.select do |activity|
-        activity.visible_for?(current_user)
-      end
+      @activities ||= last_activities
     end
 
     def last_activities
       @last_activities ||= model.map do |activity|
+        next unless activity.visible_for?(current_user)
+
         activity.organization_lazy
         activity.resource_lazy
         activity.participatory_space_lazy