#### :tophat: What? Why?
Just a little optimization on last activities loading. 

I noticed an extra loop that we can refactor 

**Benchmarks**

Loading on development mode : 

Before: 
```
  Rendered cell decidim/comments/comment_activity (144.3ms)
  Rendered cell decidim/comments/comment_activity (101.8ms)
  Rendered cell decidim/comments/comment_activity (84.5ms)
  Rendered cell decidim/comments/comment_activity (83.9ms)
  Rendered cell decidim/comments/comment_activity (77.1ms)
  Rendered cell decidim/comments/comment_activity (76.5ms)
  Rendered cell decidim/comments/comment_activity (87.3ms)
  Rendered cell decidim/comments/comment_activity (73.1ms)
```

After:
```
  Rendered cell decidim/comments/comment_activity (107.2ms)
  Rendered cell decidim/comments/comment_activity (92.2ms)
  Rendered cell decidim/comments/comment_activity (70.2ms)
  Rendered cell decidim/comments/comment_activity (72.5ms)
  Rendered cell decidim/comments/comment_activity (72.2ms)
  Rendered cell decidim/comments/comment_activity (63.1ms)
  Rendered cell decidim/comments/comment_activity (76.6ms)
  Rendered cell decidim/comments/comment_activity (65.8ms)
```
#### :pushpin: Related Issues
- No related issue, just a light refactor

:hearts: Thank you!