@Quentinchampenois 2022-10-19
I am closing this draft PR because this optimization is not significative